branchmap: create a mercurial.branchmap module
This is the foundation stone for an extraction of branches map logic from local
repository class.  Most of the branches map logic have very few caller and
therefor does not fit in the current criteria for code held by the localrepo
class. Important change will be made to this code in relation with revision
filtering. So we extract it in a dedicated module before adding additional
complexity.


Follow up commit do the actual code movement.